requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==24.2.0
backports.zoneinfo==0.2.1
billiard==4.2.0
black==24.8.0
celery==5.4.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.1
Django==4.2.15
django-filter==24.3
django-oauth-toolkit==3.0.1
django-rest-swagger==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
exceptiongroup==1.2.2
h11==0.14.0
idna==3.7
inflection==0.5.1
itypes==1.2.0
jinja2==3.1.4
jwcrypto==1.5.6
kombu==5.4.1
MarkupSafe==2.1.5
mypy-extensions==1.0.0
mysqlclient==2.2.4
oauthlib==3.2.2
openapi-codec==1.3.2
outcome==1.3.0.post0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
prompt-toolkit==3.0.47
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.9.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
selenium==4.23.1
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
tomli==2.0.1
trio==0.26.2
trio-websocket==0.11.1
typing-extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
